Fixed glitch in read_cooler/bins where bin count was read as integer, potentially giving an overflow error when calculating number of bins to retrieve.

+++ R/read_cooler_bins_hdf5.R
@@ -37,7 +37,7 @@ read_cooler_bins_hdf5 <- function(file_cooler,granges_list = NULL,ignore_strand=
   H5Fclose(hdf5)
   gr_bins <- bin_data %>%
     as_tibble %>%
-    mutate(bin_id = as.integer(row_number()-1),
+    mutate(bin_id = as.double(row_number()-1),
            start_adj = start + sq_offs[as.character(chrom)],
            end_adj = end + sq_offs[as.character(chrom)]) %>%
     mutate(chrom = factor(chrom,levels=paste0("chr",c(1:22,"X","Y")))) %>%
